+++ utilities/pycvm/pycvm/horizontal_slice.py
@@ -172,7 +172,7 @@ class HorizontalSlice:
         lats = np.linspace(self.bottomrightpoint.latitude, self.upperleftpoint.latitude - self.spacing, self.num_y-1)
     
         # Get the properties.
-        datapoints = np.arange(self.num_x * self.num_y,dtype=float).reshape(self.num_y, self.num_x)
+        datapoints = np.arange(self.num_x * self.num_y,dtype=np.float32).reshape(self.num_y, self.num_x)
 
         nancnt=0
         zerocnt=0
